# src/modules/access-control/access-control.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Put,
  Param,
  Query,
  UseGuards,
} from '@nestjs/common';
import { AccessControlService } from './access-control.service';
import { JwtAuthGuard } from '../auth/jwt-auth.guard';
import { AbilitiesGuard } from './abilities.guard';
import { CheckAbilities } from './abilities.decorator';
import { Action } from './ability.factory';

@Controller('access-control')
@UseGuards(JwtAuthGuard)
export class AccessControlController {
  constructor(private readonly accessControlService: AccessControlService) {}

  @Get('roles')
  async findAllRoles(@Query('companyId') companyId: string) {
    return this.accessControlService.findAllRoles(companyId);
  }

  @Get('roles/:id')
  async findOneRole(@Param('id') id: string) {
    return this.accessControlService.findOneRole(id);
  }

  @Post('roles')
  async createRole(@Body() data: any) {
    return this.accessControlService.createRole(data);
  }

  @Put('roles/:id')
  async updateRole(@Param('id') id: string, @Body() data: any) {
    return this.accessControlService.updateRole(id, data);
  }

  @Get('permissions')
  async findAllPermissions() {
    return this.accessControlService.findAllPermissions();
  }

  @Post('permissions')
  async createPermission(@Body() data: any) {
    return this.accessControlService.createPermission(data);
  }

  @Post('seed')
  async seed() {
    return this.accessControlService.seedDefaultPermissions();
  }
}