# swagger.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - App
  /dashboard/metrics:
    get:
      operationId: DashboardController_getMetrics
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Dashboard
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      responses:
        "201":
          description: ""
      tags: &a1
        - Auth
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a1
  /auth/profile:
    get:
      operationId: AuthController_getProfile
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a1
    patch:
      operationId: AuthController_updateProfile
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a1
  /company:
    post:
      operationId: CompanyController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyDto"
      responses:
        "201":
          description: ""
      tags: &a2
        - Company
  /company/id/{id}:
    get:
      operationId: CompanyController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a2
  /company/{email}:
    get:
      operationId: CompanyController_findOne
      parameters:
        - name: email
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a2
  /company/{id}:
    patch:
      operationId: CompanyController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a2
  /departments:
    post:
      operationId: DepartmentController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentDto"
      responses:
        "201":
          description: ""
      tags: &a3
        - Department
    get:
      operationId: DepartmentController_findAll
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a3
  /departments/{id}:
    get:
      operationId: DepartmentController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a3
    patch:
      operationId: DepartmentController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentDto"
      responses:
        "200":
          description: ""
      tags: *a3
    delete:
      operationId: DepartmentController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a3
  /access-control/roles:
    get:
      operationId: AccessControlController_findAllRoles
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: &a4
        - AccessControl
    post:
      operationId: AccessControlController_createRole
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a4
  /access-control/roles/{id}:
    get:
      operationId: AccessControlController_findOneRole
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a4
    put:
      operationId: AccessControlController_updateRole
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a4
  /access-control/permissions:
    get:
      operationId: AccessControlController_findAllPermissions
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a4
    post:
      operationId: AccessControlController_createPermission
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a4
  /access-control/seed:
    post:
      operationId: AccessControlController_seed
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a4
  /employees:
    post:
      operationId: EmployeeController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeDto"
      responses:
        "201":
          description: ""
      tags: &a5
        - Employee
    get:
      operationId: EmployeeController_findAll
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a5
  /employees/fix-roles:
    post:
      operationId: EmployeeController_fixRoles
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a5
  /employees/stats/count:
    get:
      operationId: EmployeeController_getCount
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a5
  /employees/{id}:
    get:
      operationId: EmployeeController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a5
    put:
      operationId: EmployeeController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeDto"
      responses:
        "200":
          description: ""
      tags: *a5
    delete:
      operationId: EmployeeController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a5
  /attendance/check-in:
    post:
      operationId: AttendanceController_checkIn
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInDto"
      responses:
        "201":
          description: ""
      tags: &a6
        - Attendance
  /attendance/bulk-check-in:
    post:
      operationId: AttendanceController_bulkCheckIn
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCheckInDto"
      responses:
        "201":
          description: ""
      tags: *a6
  /attendance/check-out:
    post:
      operationId: AttendanceController_checkOut
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckOutDto"
      responses:
        "201":
          description: ""
      tags: *a6
  /attendance:
    post:
      operationId: AttendanceController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceDto"
      responses:
        "201":
          description: ""
      tags: *a6
    get:
      operationId: AttendanceController_findAll
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a6
  /attendance/today/{employeeId}:
    get:
      operationId: AttendanceController_getTodayAttendance
      parameters:
        - name: employeeId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a6
  /attendance/analytics:
    get:
      operationId: AttendanceController_getAnalytics
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a6
  /attendance/{id}:
    get:
      operationId: AttendanceController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a6
  /leaves:
    post:
      operationId: LeavesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveDto"
      responses:
        "201":
          description: ""
      tags: &a7
        - Leaves
    get:
      operationId: LeavesController_findAll
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a7
  /leaves/{id}:
    get:
      operationId: LeavesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a7
  /leaves/{id}/status:
    patch:
      operationId: LeavesController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveStatusDto"
      responses:
        "200":
          description: ""
      tags: *a7
  /designations:
    post:
      operationId: DesignationController_create
      parameters: []
      responses:
        "201":
          description: ""
      tags: &a8
        - Designation
    get:
      operationId: DesignationController_findAll
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a8
  /designations/{id}:
    get:
      operationId: DesignationController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a8
    put:
      operationId: DesignationController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a8
    delete:
      operationId: DesignationController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a8
  /projects/clients:
    post:
      operationId: ProjectsController_createClient
      parameters: []
      responses:
        "201":
          description: ""
      tags: &a9
        - Projects
    get:
      operationId: ProjectsController_findAllClients
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a9
  /projects:
    post:
      operationId: ProjectsController_createProject
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a9
    get:
      operationId: ProjectsController_findAllProjects
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a9
  /projects/tasks:
    post:
      operationId: ProjectsController_createTask
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a9
    get:
      operationId: ProjectsController_findAllTasks
      parameters:
        - name: projectId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a9
  /projects/tasks/company:
    get:
      operationId: ProjectsController_findAllTasksByCompany
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a9
  /projects/tasks/{id}:
    put:
      operationId: ProjectsController_updateTask
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a9
  /finance/payroll:
    post:
      operationId: FinanceController_createPayroll
      parameters: []
      responses:
        "201":
          description: ""
      tags: &a10
        - Finance
    get:
      operationId: FinanceController_findAllPayroll
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
        - name: month
          required: true
          in: query
          schema:
            type: number
        - name: year
          required: true
          in: query
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags: *a10
  /finance/expenses:
    post:
      operationId: FinanceController_createExpense
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a10
    get:
      operationId: FinanceController_findAllExpenses
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a10
  /finance/summary:
    get:
      operationId: FinanceController_getSummary
      parameters:
        - name: companyId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a10
  /time-tracking/start:
    post:
      operationId: TimeTrackingController_start
      parameters: []
      responses:
        "201":
          description: ""
      tags: &a11
        - TimeTracking
  /time-tracking/stop/{id}:
    put:
      operationId: TimeTrackingController_stop
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a11
  /time-tracking:
    get:
      operationId: TimeTrackingController_findAll
      parameters:
        - name: employeeId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a11
  /add-projects:
    post:
      operationId: ProjectController_create
      parameters: []
      responses:
        "201":
          description: ""
      tags: &a12
        - Project
    get:
      operationId: ProjectController_findAll
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a12
  /add-projects/{id}:
    get:
      operationId: ProjectController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a12
    put:
      operationId: ProjectController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a12
    delete:
      operationId: ProjectController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a12
info:
  title: Difmo CRM API
  description: The Difmo CRM API description
  version: "1.0"
  contact: {}
tags: []
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://api.difmocrm.com
    description: Production
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateCompanyDto:
      type: object
      properties: {}
    CreateDepartmentDto:
      type: object
      properties: {}
    UpdateDepartmentDto:
      type: object
      properties: {}
    CreateEmployeeDto:
      type: object
      properties: {}
    UpdateEmployeeDto:
      type: object
      properties: {}
    CheckInDto:
      type: object
      properties: {}
    BulkCheckInDto:
      type: object
      properties: {}
    CheckOutDto:
      type: object
      properties: {}
    CreateAttendanceDto:
      type: object
      properties: {}
    CreateLeaveDto:
      type: object
      properties: {}
    UpdateLeaveStatusDto:
      type: object
      properties: {}